requirements.txt
pandas
streamlit
pymongo
psycopg2-binary
python-dotenv
mysql-connector-python
python-dateutil
